# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: CrabIndex API
  version: 0.1.0 # replaced at runtime with the crabindex crate version
  description: |
    Torrent aggregator API: Jackett-compatible search, native torrent search, Torznab XML,
    statistics and sync protocol.

    **Interactive docs:** [Swagger UI](/swagger)

    **Runtime OpenAPI JSON:** `/swagger/v1/swagger.json` (converted from this YAML file)

    **Static spec (this file):** `/openapi.yaml` (source: `web/public/openapi.yaml`, published to `wwwroot/openapi.yaml`)

    ## Supported tracker slugs

    Parsers / cron (`/cron/{slug}/…`, DevAdmin):
    `anibelka`, `anidub`, `anifilm`, `aniliberty`, `animelayer`, `anistar`, `baibako`,
    `bitru`, `kinozal`, `knaben`, `korsars`, `leproduction`, `lostfilm`, `mazepa`,
    `megapeer`, `nnmclub`, `rudub`, `rutor`, `rutracker`, `selezen`, `subsplease`, `toloka`, `torrentby`,
    `ultradox`, `viruseproject`.

    Web UI icons: `/img/ico/{slug}.ico` (fallback `/img/ico/default.ico`).

    Page-only parsers (`anistar`, `leproduction`, `viruseproject`, `anifilm`) usually do **not**
    appear in `GET /health/background-jobs` (only UpdateTasks/ParseAll-style workers register there).

    ## Authentication

    | Scheme | Usage |
    |--------|--------|
    | `apikey` query | Search, stats (when `apikey` is set in config) |
    | `X-Api-Key` header | Same as query apikey |
    | `Authorization: Bearer` | Same as apikey |
...
    | `devkey` query | Same as `X-Dev-Key` (`?devkey=`) |

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v2.0/indexers/{status}/results:
    get:
      tags: [Search]
      summary: Jackett-compatible search
      description: |
        Lampa and other Jackett JSON clients. Path `{status}` is normally `all` (aggregate).
        Jackett's virtual `status:healthy` selector (Lampa "only available trackers") is also
        treated as aggregate - CrabIndex searches its local database, not live indexer health.
        When `{status}` is a tracker id from `GET /api/v1.0/trackers` (for example `rutracker`),
        results are filtered to that tracker (case-insensitive; matches any part of a merged trackerName).

        **IMDb / Kinopoisk / TMDB IDs:** `query`/`q` as `tt…`, `kp…`, `tmdb…`, a themoviedb.org
        `/movie|tv/{id}-…` URL, or `imdbid`, are resolved via Alloha TV API v2
        (`GET /v2/movies/search`) to titles (plus optional alternative name, year, type), then searched
        in FileDB with exact match. When the client omits `year`, Alloha year ±1 is applied if
        `alloha.filterByYear` is enabled.

        **TV text queries:** when `search.stripSeasonEpisode` is true (default), inline `S01` / `S01E01`
        tokens are stripped and both variants are searched (AIOStreams-style `q=silo S01`).
        Trailing year is also stripped when `search.stripTrailingYear` is true.
      operationId: jackettSearch
      parameters:
        - name: status
          in: path
          required: true
          description: |
            `all` or Jackett `status:healthy` for aggregate search, or a tracker id
            (e.g. `kinozal`, `rutracker`) to scope results.
          schema:
            type: string
            example: all
        - $ref: "#/components/parameters/ApiKeyQuery"
        - name: query
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/SearchQ"
        - $ref: "#/components/parameters/SearchQueryAlt"
        - $ref: "#/components/parameters/TorznabTitle"
        - $ref: "#/components/parameters/TorznabTitleOriginal"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/IsSerial"
        - $ref: "#/components/parameters/CategoryBracket"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/TorznabCat"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/TrackerArray"
        - $ref: "#/components/parameters/TrackerFilter"
        - $ref: "#/components/parameters/TrackerAlias"
        - $ref: "#/components/parameters/TorznabImdbId"
        - $ref: "#/components/parameters/ImdbIdAltUnderscore"
        - $ref: "#/components/parameters/ImdbIdAltCamel"
        - $ref: "#/components/parameters/TorznabSeason"
        - $ref: "#/components/parameters/TorznabEp"
        - $ref: "#/components/parameters/TorznabEpisode"
        - $ref: "#/components/parameters/TorznabLimit"
        - $ref: "#/components/parameters/TorznabOffset"
        - $ref: "#/components/parameters/Genres"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JackettRootObject"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1.0/torrents:
    get:
      tags: [Search]
      summary: Native torrent search
      operationId: torrentsSearch
      description: |
        Native CrabIndex JSON search (not Torznab/Jackett).

        **IMDb / Kinopoisk / TMDB IDs:** when `search` is `tt…`, `kp…`, `tmdb…`, or a themoviedb.org
        `/movie|tv/{id}-…` URL, CrabIndex resolves titles via Alloha TV API v2,
        forces `exact=true`, and may fill `type` / year filter from Alloha when the client omitted them
        (`relased` unset → Alloha year ±1 if `alloha.filterByYear`; empty `type` → Alloha category).
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: |
            Title query, or an external id (`tt1234567`, `kp301`, `tmdb1315772`) or themoviedb.org URL.
            Empty or omitted yields an empty result array.
            ID queries are resolved via Alloha v2 before FileDB lookup.
        - name: altname
          in: query
          schema:
            type: string
          description: Alternate title (e.g. Russian name). After Alloha resolve, filled from `data.name` when both titles exist.
        - name: exact
          in: query
          schema:
            type: boolean
          description: Exact title match. Forced to true after IMDb/KP Alloha resolve.
        - name: type
          in: query
          schema:
            type: string
            description: |
              movie, serial, anime, etc. When omitted on an ID query, may be set from Alloha `category.slug`.
        - name: sort
          in: query
          schema:
            type: string
            enum: [sid, pir, size, create, update]
        - $ref: "#/components/parameters/TrackerAlias"
        - name: voice
          in: query
          schema:
            type: string
        - name: videotype
          in: query
          schema:
            type: string
        - name: relased
          in: query
          schema:
            type: integer
          description: |
            Release year filter (exact). When 0/omitted on an ID query and `alloha.filterByYear`,
            Alloha year ±1 is used instead.
        - name: quality
          in: query
          schema:
            type: integer
        - name: season
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ordered torrent result array (maximum 2000 items)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TorrentListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1.0/trackers:
    get:
      tags: [Search]
      summary: List available tracker names
      description: |
        Returns tracker ids for discovery and filters.
        Source: `synctrackers` when set (including empty `[]` → empty list); otherwise known built-in slugs.
        Entries in `disable_trackers` are excluded. Sorted case-insensitively.
      operationId: trackerNames
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Available tracker names (e.g. rutracker, kinozal, korsars)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackerSlug"
                example: [kinozal, korsars, rutracker, rutor]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1.0/qualitys:
    get:
      tags: [Search]
      summary: Quality aggregation by title
      operationId: qualitys
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
        - name: name
          in: query
          schema:
            type: string
        - name: originalname
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: take
          in: query
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: |
            Nested map: title key → release year → TorrentQuality.
            Empty name and originalname yields `{}`.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    $ref: "#/components/schemas/TorrentQuality"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /torznab/api:
    get:
      tags: [Torznab]
      summary: Torznab API (caps, search)
      description: |
        Native Torznab XML when `torznab.enable` is true.
        Caps advertise `q`, `imdbid`[, `tvdbid`, `season`, `ep`]. CrabIndex also accepts card fields
        (`title`, `title_original`, `year`, `is_serial`, `genres`), pagination (`limit`/`offset`),
        category aliases (`cat`/`Category`/`categories`/`Category[]`), and tracker filters
        (`Tracker`/`Tracker[]`/`tracker`).
        `t=indexers` returns aggregate `all` plus one indexer entry per `GET /api/v1.0/trackers` name.

        **IMDb / Kinopoisk / TMDB IDs:** `q` as `tt…`/`kp…`/`tmdb…`, a themoviedb.org URL, or `imdbid` →
        Alloha TV API v2 title resolve → FileDB exact search (year ±1 when client `year` absent and `alloha.filterByYear`).
        **TVDB-only** (`tvdbid` / `rid` with no `q`/`imdbid`) returns empty - CrabIndex does not resolve TVDB ids.
        **TV text queries:** `search.stripSeasonEpisode` (default true) searches show name without `S01`/`S01E01`;
        `season`/`ep` still post-filter unless `search.skipSeasonEpisodeFilter` is true.
      operationId: torznabRoot
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
        - $ref: "#/components/parameters/TorznabT"
        - $ref: "#/components/parameters/TorznabQ"
        - $ref: "#/components/parameters/SearchQueryAlt"
        - $ref: "#/components/parameters/TorznabCat"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/CategoryBracket"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/TorznabImdbId"
        - $ref: "#/components/parameters/ImdbIdAltUnderscore"
        - $ref: "#/components/parameters/ImdbIdAltCamel"
        - $ref: "#/components/parameters/TorznabTvdbId"
        - $ref: "#/components/parameters/TorznabRid"
        - $ref: "#/components/parameters/TorznabSeason"
        - $ref: "#/components/parameters/TorznabEp"
        - $ref: "#/components/parameters/TorznabEpisode"
        - $ref: "#/components/parameters/TorznabTitle"
        - $ref: "#/components/parameters/TorznabTitleOriginal"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/IsSerial"
        - $ref: "#/components/parameters/Genres"
        - $ref: "#/components/parameters/TorznabLimit"
        - $ref: "#/components/parameters/TorznabOffset"
        - $ref: "#/components/parameters/TorznabConfigured"
        - $ref: "#/components/parameters/TrackerFilter"
        - $ref: "#/components/parameters/TrackerArray"
        - $ref: "#/components/parameters/TrackerAlias"
      responses:
        "200":
          description: Torznab XML
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
    head:
      tags: [Torznab]
      summary: Torznab HEAD probe
      description: Returns empty Torznab XML body (same content-type as GET).
      operationId: torznabRootHead
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Empty XML OK
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
  /api/v2.0/indexers/{indexer}/results/torznab/api:
    get:
      tags: [Torznab]
      summary: Torznab API (indexer path)
      description: |
        Jackett-compatible Torznab alias. Path `{indexer}` is normally `all`.
        Jackett's virtual `status:healthy` selector is also treated as aggregate.
        When `{indexer}` is a tracker id (not `all` / `status:healthy`), search results are filtered to that tracker.
        `t=indexers` lists `all` plus per-tracker indexers from the tracker catalog.
        Same query params as `/torznab/api` (caps + CrabIndex card fields, categories, tracker filters),
        including Alloha v2 resolve for `tt…` / `kp…` / `tmdb…` / themoviedb.org URLs / `imdbid`.
      operationId: torznabIndexer
      parameters:
        - name: indexer
          in: path
          required: true
          description: |
            `all` or Jackett `status:healthy` for aggregate search, or a tracker id
            (e.g. `rutracker`) to scope results.
          schema:
            type: string
            example: all
        - $ref: "#/components/parameters/ApiKeyQuery"
        - $ref: "#/components/parameters/TorznabT"
        - $ref: "#/components/parameters/TorznabQ"
        - $ref: "#/components/parameters/SearchQueryAlt"
        - $ref: "#/components/parameters/TorznabCat"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/CategoryBracket"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/TorznabImdbId"
        - $ref: "#/components/parameters/ImdbIdAltUnderscore"
        - $ref: "#/components/parameters/ImdbIdAltCamel"
        - $ref: "#/components/parameters/TorznabTvdbId"
        - $ref: "#/components/parameters/TorznabRid"
        - $ref: "#/components/parameters/TorznabSeason"
        - $ref: "#/components/parameters/TorznabEp"
        - $ref: "#/components/parameters/TorznabEpisode"
        - $ref: "#/components/parameters/TorznabTitle"
        - $ref: "#/components/parameters/TorznabTitleOriginal"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/IsSerial"
        - $ref: "#/components/parameters/Genres"
        - $ref: "#/components/parameters/TorznabLimit"
        - $ref: "#/components/parameters/TorznabOffset"
        - $ref: "#/components/parameters/TorznabConfigured"
        - $ref: "#/components/parameters/TrackerFilter"
        - $ref: "#/components/parameters/TrackerArray"
        - $ref: "#/components/parameters/TrackerAlias"
      responses:
        "200":
          description: Torznab XML
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
    head:
      tags: [Torznab]
      summary: Torznab HEAD probe (indexer path)
      description: Returns empty Torznab XML body (same content-type as GET).
      operationId: torznabIndexerHead
      parameters:
        - name: indexer
          in: path
          required: true
          schema:
            type: string
            example: all
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Empty XML OK
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
  /api/v2.0/indexers:
    get:
      tags: [Torznab]
      summary: List configured indexers (Jackett)
      description: |
        Jackett-style indexer discovery. Always includes aggregate `id=all`, then one entry per
        `GET /api/v1.0/trackers` name (same set as Torznab `t=indexers`).
        Prowlarr clients should use `/api/v1/indexer` (single aggregate `id=1`).
      operationId: indexersList
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Indexer list (`all` + per-tracker)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JackettIndexer"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/indexer:
    get:
      tags: [Torznab]
      summary: Prowlarr indexer list
      description: |
        Prowlarr-compatible indexer list for discover clients (qui, autobrr) and Prowlarr apps.
        Returns a JSON array with `enable: true` and `protocol: torrent`.
      operationId: prowlarrIndexerList
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
        - name: X-Api-Key
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Enabled torrent indexers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProwlarrIndexer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
  /api/v1/indexer/{id}:
    get:
      tags: [Torznab]
      summary: Prowlarr indexer detail
      description: |
        Prowlarr-compatible indexer detail for qui tracker domain resolution (backend=prowlarr).
        CrabIndex exposes a single aggregate indexer with `id=1`.
      operationId: prowlarrIndexerDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Indexer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProwlarrIndexerDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled or unknown indexer id
...
components:
  securitySchemes:
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apikey
      description: API key (when configured in init.yaml)
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key as Bearer token
    DevKeyHeader:
      type: apiKey
      in: header
      name: X-Dev-Key
      description: Admin/dev key - required from internet or tunnel when `devkey` is set; LAN clients may omit it.
    DevKeyQuery:
      type: apiKey
      in: query
      name: devkey
      description: Same as X-Dev-Key (`?devkey=`).
  parameters:
    ApiKeyQuery:
      name: apikey
      in: query
      required: false
      schema:
        type: string
      description: API key (alternative - X-Api-Key header or Bearer)
...
    TorznabTitle:
      name: title
      in: query
      schema:
        type: string
      description: Localized title (CrabIndex / Lampa card search)
    TorznabTitleOriginal:
      name: title_original
      in: query
      schema:
        type: string
      description: Original title (CrabIndex / Lampa card search)
...
    SearchQ:
      name: q
      in: query
      schema:
        type: string
      description: |
        Free-text search alias for `query` / Torznab `q`. Same season/year stripping as Torznab `q`.
    SearchQueryAlt:
      name: Query
      in: query
      schema:
        type: string
      description: Alternate capitalization of free-text search query (same stripping as `q`)
    ImdbIdAltUnderscore:
      name: imdb_id
      in: query
      schema:
        type: string
      description: Alias for `imdbid`
...
    Year:
      name: year
      in: query
      schema:
        type: integer
      description: |
        Release year filter / card-search override.
        When omitted on an IMDb/KP/TMDB id query and `alloha.filterByYear` is enabled, Alloha year ±1 is applied.
...
security:
  - ApiKeyQuery: []
  - ApiKeyHeader: []
  - Bearer: []